+++ presentation/src/main/kotlin/com/mickeycj/restcountries/main/MainActivity.kt
@@ -1,10 +1,14 @@
-package com.mickeycj.restcountries
+package com.mickeycj.restcountries.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.mickeycj.restcountries.R
 
 import kotlinx.android.synthetic.main.activity_main.*
 
+/**
+ * Activity for Main screens.
+ */
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
