+++ presentation/src/main/kotlin/com/mickeycj/restcountries/main/countries/CountriesFragment.kt
@@ -1,15 +1,19 @@
-package com.mickeycj.restcountries
+package com.mickeycj.restcountries.main.countries
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.mickeycj.restcountries.R
 
-class CountryDetailsFragment : Fragment() {
+/**
+ * Fragment for Countries screen.
+ */
+class CountriesFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_country_details, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_countries, container, false)
 }
