+++ presentation/src/main/kotlin/com/mickeycj/restcountries/main/countrydetails/CountryDetailsFragment.kt
@@ -1,15 +1,19 @@
-package com.mickeycj.restcountries
+package com.mickeycj.restcountries.main.countrydetails
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.mickeycj.restcountries.R
 
-class CountriesFragment : Fragment() {
+/**
+ * Fragment for Country Details screen.
+ */
+class CountryDetailsFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_countries, container, false)
+    ): View? = inflater.inflate(R.layout.fragment_country_details, container, false)
 }
