+++ presentation/src/main/kotlin/com/mickeycj/restcountries/splash/SplashActivity.kt
@@ -1,8 +1,12 @@
-package com.mickeycj.restcountries
+package com.mickeycj.restcountries.splash
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.mickeycj.restcountries.R
 
+/**
+ * Activity for Splash screen.
+ */
 class SplashActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
