+++ presentation/src/main/res/layout/activity_splash.xml
@@ -6,7 +6,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:gravity="center"
-        tools:context=".SplashActivity">
+        tools:context=".splash.SplashActivity">
 
     <de.hdodenhof.circleimageview.CircleImageView
             android:layout_width="120dp"
