+++ presentation/src/main/res/layout/fragment_country_details.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CountryDetailsFragment">
+        tools:context=".main.countrydetails.CountryDetailsFragment">
 
     <LinearLayout
             android:layout_width="match_parent"
